third_party/blink/web_tests/http/tests/security/cross-origin-access-over-property-descriptor-expected.txt

Cross-origin access through 'get' and 'set' in a property descriptor should throw a SecurityError.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS [object HTMLIFrameElement] is non-null.
PASS [object Window] is non-null.
PASS pathnameDescriptor is non-null.
PASS typeof pathnameDescriptor.get is "function"
PASS typeof pathnameDescriptor.set is "function"
PASS devicePixelRatio is non-null.
PASS typeof devicePixelRatioDescriptor.get is "function"
PASS typeof devicePixelRatioDescriptor.set is "function"
PASS pathnameDescriptor.get.call(targetWindow.location) threw exception SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame..
PASS pathnameDescriptor.set.call(targetWindow.location, 1) threw exception SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame..
PASS devicePixelRatioDescriptor.get.call(targetWindow) threw exception SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame..
PASS devicePixelRatioDescriptor.set.call(targetWindow, 1) threw exception SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame..
PASS successfullyParsed is true

TEST COMPLETE
